Guessing almost works
Still a (WIP), need to figure out how to set the text of the generated TextView to be what the user guesses correctly, but it crashes on inpact of attempting to set the text. using an array to keep track of generated TextViews, indexing properly, but still crashes. Will look into.